Configure Ruby version and update Gemfile.lock
This commit changes the specified Ruby version 3.0.0 ->  3.1.4 for the
sake of convenience, since this is what exists on my computer.

Gemfile.lock also appears to be outdated here, since the `bundle`
command failed with a reference to "Your bundle is locked to mimemagic
(0.3.5), but that version could not be found...". Since this dependency
is no longer present in rails 6.1.3 (the version used in this repo),
Gemfile.lock was removed and `bundle` re-run, resulting in several gem
upgrades.

With these changes, we can boot the server successfully.