Add page for submitting a new review
This commit adds a new route, linked to a new `.slim` file, for creating
a new review.

I decided to utilize two select elements to allow the user to choose the
shop and product; once they have chosen a shop, the products for said
shop will be fetched and populated in the second select element.

Styling to be added later.