Refactor index action
This commit refactors the index actions in two key ways.

1. Handles case where no shop ID is provided. Since the only differece
here is a filter by shop ID in the event one is provided, we can
conditionally add this to the query.

2. Addresses N+1 issue: The prior way of fetching reviews caused an N+1
issue, as the `reviews` table was being queried for each iteration of
`products.map`. Instead, we grab all relevant reviews beforehand, and
simply build the relevant return object.